requirements.txt
angr==9.2.41
# binwalk==2.3.3+a555eb1
chardet==5.2.0
capstone==4.0.2

lxml==4.9.2
networkx==3.0
openpyxl==3.1.2
psycopg2-binary==2.9.5
PyYAML==6.0
rich==13.3.2

numpy==1.24.2
scipy==1.10.1

black==23.3.0

pytest==7.1.2
pytest-xdist==3.2.1

gprof2dot==2022.7.29
graphviz==0.20.1
ipython==8.11.0
requests